Separate auth into register and login routes
