+++ client/src/routes/register/+page.server.ts
@@ -1,53 +1,10 @@
-// src/routes/+page.server.ts
+
 import { fail, redirect } from '@sveltejs/kit';
 import type { Actions } from './$types';
 
 export const actions = {
-    login: async ({ request, cookies }) => {
-        const data = await request.formData();
-        const email = data.get('email')?.toString() ?? '';
-        const password = data.get('password')?.toString() ?? '';
-
-        const formValues = { email }; // We won't include password for security reasons
-
-        if (!email || !password) {
-            return fail(400, {
-                ...formValues,
-                errors: { general: 'Email and password are required.' }
-            });
-        }
-
-        try {
-            const response = await fetch('http://localhost:5072/account/login', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ email, password }),
-                credentials: 'include'
-            });
-
-            if (!response.ok) {
-                const errorData = await response.json();
-                return fail(response.status, {
-                    ...formValues,
-                    errors: { general: errorData.message }
-                });
-            }
-
-            // Handle successful login (e.g., set cookies if necessary)
-            // cookies.set('sessionid', '...', { path: '/' });
-        } catch (error) {
-            console.error(error);
-            return fail(500, {
-                ...formValues,
-                errors: { general: 'Server error during login.' }
-            });
-        }
-
-        // Redirect to the homepage after successful login
-        throw redirect(303, '/');
-    },
-
     register: async ({ request }) => {
+        console.log('before')
         const data = await request.formData();
         const name = data.get('name')?.toString() ?? '';
         const email = data.get('email')?.toString() ?? '';
@@ -72,6 +29,7 @@ export const actions = {
         }
 
         if (Object.keys(errors).length > 0) {
+            console.log('throwing')
             return fail(400, {
                 ...formValues,
                 errors
@@ -103,6 +61,9 @@ export const actions = {
 
             // Handle successful registration (e.g., set cookies if necessary)
             // cookies.set('sessionid', '...', { path: '/' });
+
+            // Registration and sign-in successful
+            throw redirect(303, '/');
         } catch (error) {
             console.error(error);
             return fail(500, {
@@ -110,8 +71,5 @@ export const actions = {
                 errors: { general: 'Server error during registration.' }
             });
         }
-
-        // Registration and sign-in successful
-        throw redirect(303, '/');
     }
 } satisfies Actions;
